Refactored Bryan Rumsey's simulations settings
