+++ src/templates/includes/deterministicSettings.pug
@@ -8,11 +8,11 @@ div#deterministic-settings.card.card-body
         th(scope="col", colspan="2") Relative Tolerance
 
       tr
-        td(colspan="2"): div(data-hook="relative-tolerance-container") Relative tolerance of the ODE solver (CVODES).  Valid valeus must be greater than 0.0:  
+        td(colspan="2"): div(data-hook="relative-tolerance-container")
 
     thread
       tr
         th(scope="col", colspan="2") Absolute Tolerance
 
       tr
-        td(colspan="2"): div(data-hook="absolute-tolerance-container") Absolute Tolerance of the ODE solver (CVODES).  Valid values must be greater than 0.0:  
+        td(colspan="2"): div(data-hook="absolute-tolerance-container")
