+++ src/templates/includes/modelVersionEditor.pug
@@ -6,6 +6,6 @@ div#model-version-editor(data-hook="model-version-editor-container")
   
   div(data-hook="reactions-editor-container")
 
-  div(data-hook="run-model-container")
+  div(data-hook="sim-settings-container")
 
   div(data-hook="model-state-buttons-container")
