--- src/templates/includes/runModel.pug
@@ -1,17 +1,0 @@
-div#run-model.card.card-body
-
-  h3 Run Model
-
-  div(data-hook="sim-type-settings-container")
-
-  table.table
-    thread
-      tr
-        th(scope="col") Simulation Time
-
-        th(scope="col") Time Units
-
-      tr
-        td: div(data-hook="end-sim-container")
-        
-        td: div(data-hook="time-units-container")
